requirements.txt
fastapi==0.103.1
uvicorn==0.23.2
openai==1.70.0
python-dotenv==1.0.0
pydantic==2.3.0
requests==2.31.0
pytest==7.4.0
httpx==0.24.1 
